Make Reader public
pars uses a custom reader that supports unlimited unreading. This is
used for backtracking when a parser fails.
Until now, the reader could only used indirectly via the helper
functions ParseString and ParseFromReader (which takes an io.Reader).
Both functions only allow for a single parser. If that parser fails, the
io.Reader is not unread (because it cannot be) and the next call to
ParseFromReader will continue after the partially read position.

Using pars.NewReader directly allows for better input streaming as
unreading works across multiple calls.